+++ rasa_core/policies/ensemble.py
@@ -174,7 +174,7 @@ class PolicyEnsemble(object):
         from packaging import version
 
         model_version = metadata.get("rasa_core", "0.0.0")
-        if version.parse(model_version) < version.parse("0.9.0"):
+        if version.parse(model_version) < version.parse("0.10.0a3"):
             raise UnsupportedDialogueModelError(
                 "The model version is to old to be "
                 "loaded by this Rasa Core instance. "
@@ -218,7 +218,9 @@ class SimplePolicyEnsemble(PolicyEnsemble):
                 max_confidence = confidence
                 result = probabilities
                 best_policy_name = 'policy_{}_{}'.format(i, type(p).__name__)
-
+        # normalize probablilities
+        if np.sum(result) != 0:
+            result = result / np.linalg.norm(result)
         logger.debug("Predicted next action using {}"
                      "".format(best_policy_name))
         return result