+++ rasa_core/server.py
@@ -10,22 +10,19 @@ import tempfile
 import zipfile
 from functools import wraps
 
+import typing
 from builtins import str
 from flask import Flask, request, abort, Response, jsonify
 from flask_cors import CORS, cross_origin
 from gevent.pywsgi import WSGIServer
 from typing import Union, Text, Optional
 
-from rasa_core import utils, events
+from rasa_core import utils, events, run
 from rasa_core.agent import Agent
 from rasa_core.channels.direct import CollectingOutputChannel
-from rasa_core.interpreter import NaturalLanguageInterpreter
 from rasa_core.tracker_store import TrackerStore
 from rasa_core.trackers import DialogueStateTracker
 from rasa_core.version import __version__
-
-from typing import Union
-import typing
 
 if typing.TYPE_CHECKING:
     from rasa_core.interpreter import NaturalLanguageInterpreter as NLI
@@ -68,6 +65,10 @@ def create_argument_parser():
             type=str,
             default="rasa_core.log",
             help="store log file in specified file")
+    parser.add_argument(
+            '--endpoints',
+            default=None,
+            help="Configuration file for the connectors as a yml file")
 
     utils.add_logging_option_arguments(parser)
     return parser
@@ -105,7 +106,7 @@ def bool_arg(name, default=True):
 
 def request_parameters():
     if request.method == 'GET':
-        return request.args
+        return request.args.to_dict()
     else:
 
         try:
@@ -137,14 +138,16 @@ def _create_agent(
         model_directory,  # type: Text
         interpreter,  # type: Union[Text,NLI,None]
         action_factory=None,  # type: Optional[Text]
-        tracker_store=None  # type: Optional[TrackerStore]
+        tracker_store=None,  # type: Optional[TrackerStore]
+        generator=None
 ):
     # type: (...) -> Optional[Agent]
     try:
 
         return Agent.load(model_directory, interpreter,
                           tracker_store=tracker_store,
-                          action_factory=action_factory)
+                          action_factory=action_factory,
+                          generator=generator)
     except Exception as e:
         logger.warn("Failed to load any agent model. Running "
                     "Rasa Core server with out loaded model now. {}"
@@ -159,7 +162,8 @@ def create_app(model_directory,  # type: Text
                cors_origins=None,  # type: Optional[List[Text]]
                action_factory=None,  # type: Optional[Text]
                auth_token=None,  # type: Optional[Text]
-               tracker_store=None  # type: Optional[TrackerStore]
+               tracker_store=None,  # type: Optional[TrackerStore]
+               endpoints=None
                ):
     """Class representing a Rasa Core HTTP server."""
 
@@ -172,15 +176,19 @@ def create_app(model_directory,  # type: Text
         cors_origins = []
     model_directory = model_directory
 
-    interpreter = interpreter
+    nlg_endpoint = utils.read_endpoint_config(endpoints, "nlg")
+
+    nlu_endpoint = utils.read_endpoint_config(endpoints, "nlu")
 
     tracker_store = tracker_store
 
     action_factory = action_factory
 
+    _interpreter = run.interpreter_from_args(interpreter, nlu_endpoint)
+
     # this needs to be an array, so we can modify it in the nested functions...
-    _agent = [_create_agent(model_directory, interpreter,
-                            action_factory, tracker_store)]
+    _agent = [_create_agent(model_directory, _interpreter,
+                            action_factory, tracker_store, nlg_endpoint)]
 
     def agent():
         if _agent and _agent[0]:
@@ -305,6 +313,30 @@ def create_app(model_directory,  # type: Text
         agent().tracker_store.save(tracker)
         return jsonify(tracker.current_state(should_include_events=True))
 
+    @app.route("/domain",
+               methods=['GET'])
+    @cross_origin(origins=cors_origins)
+    @requires_auth(auth_token)
+    @ensure_loaded_agent(agent)
+    def get_domain():
+        """Get current domain in yaml format."""
+        accepts = request.headers.get("Accept", default="application/json")
+        if accepts.endswith("json"):
+            domain = agent().domain.as_dict()
+            return jsonify(domain)
+        elif accepts.endswith("yml"):
+            domain_yaml = agent().domain.as_yaml()
+            return Response(domain_yaml,
+                            status=200,
+                            content_type="application/x-yml")
+        else:
+            return Response(
+                    """Invalid accept header. Domain can be provided 
+                    as json ("Accept: application/json")  
+                    or yml ("Accept: application/x-yml"). 
+                    Make sure you've set the appropriate Accept header.""",
+                    status=406)
+
     @app.route("/conversations/<sender_id>/parse",
                methods=['GET', 'POST', 'OPTIONS'])
     @cross_origin(origins=cors_origins)
@@ -396,7 +428,7 @@ def create_app(model_directory,  # type: Text
                 os.path.abspath(model_directory)))
 
         _agent[0] = _create_agent(model_directory, interpreter,
-                                  action_factory, tracker_store)
+                                  action_factory, tracker_store, nlg_endpoint)
         logger.debug("Finished loading new agent.")
         return jsonify({'success': 1})
 
@@ -417,7 +449,8 @@ if __name__ == '__main__':
                      cmdline_args.loglevel,
                      cmdline_args.log_file,
                      cmdline_args.cors,
-                     auth_token=cmdline_args.auth_token)
+                     auth_token=cmdline_args.auth_token,
+                     endpoints=cmdline_args.endpoints)
 
     logger.info("Started http server on port %s" % cmdline_args.port)