+++ tests/test_remote.py
@@ -3,7 +3,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from rasa_core.actions.action import ActionListen, ActionRestart
+from rasa_core.actions.action import (ActionListen, ActionRestart,
+                                      ActionDefaultFallback)
 
 from rasa_core.actions.factories import RemoteAction
 
@@ -24,4 +25,5 @@ def test_remote_training(tmpdir):
     assert agent.domain._factory_name == "remote"
 
     action_types = [type(a) for a in agent.domain.actions]
-    assert action_types[:3] == [ActionListen, ActionRestart, RemoteAction]
+    assert action_types[:4] == [ActionListen, ActionRestart,
+                                ActionDefaultFallback, RemoteAction]